chore(utils): moved from /src to /scripts load_filter
